+++ Install.Scripts/BuildAndRunSamples/BuildAndRun.ControlsTest.sh
@@ -10,7 +10,7 @@ rm -rf obj/Release/*.*
 rm -rf obj/*.*
 
 dotnet build --property WarningLevel=0
-dotnet run --property WarningLevel=0 --framework net6.0
+dotnet run --property WarningLevel=0 --framework net8.0
 popd