+++ Source/Samples/ControlsSampleDll/Pages/InternalSamplesPage.cs
@@ -66,7 +66,10 @@ namespace ControlsSample
 
         private void AddDefaultItems()
         {
-            Add("NinePatch Drawing Sample", () => new NinePatchDrawingWindow());
+            /*
+                Add("NinePatch Drawing Sample", () => new NinePatchDrawingWindow());
+            */
+
             Add("Threading Sample", () => new ThreadingSample.ThreadingMainWindow());
             AddIfDebug("Test Page", () => new SkiaDrawingWindow());
             Add("Preview File Sample", () => new PreviewSample.PreviewSampleWindow());